manage.fabrics.fabric_delete: refactor router
1. app/main.py

- Update imports
- app.include_router(v2_fabric_delete.router, tags=["Manage Fabrics (v2)"])

2. app/v2/endpoints/fabric.py

- Remove v2_fabric_delete() endpoint handler

3, app/v2/endpoints/manage/fabrics/fabric_delete.py

- New file
- Add v2_fabric_delete() endpoint handler

4. docs/supported_endpoints.md

- Update documentation
