+++ app/v2/endpoints/fabric.py
@@ -56,32 +56,6 @@ def build_db_fabric(fabric):
     return db_fabric
 
 
-@app.delete("/api/v1/manage/fabrics/{fabric_name}", status_code=204)
-async def v2_delete_fabric(*, session: Session = Depends(get_session), fabric_name: str):
-    """
-    # Summary
-
-    DELETE request handler
-
-    ## NDFC 404 Response (4.0 LA)
-
-    ```json
-    {
-        "code": 404,
-        "description": "",
-        "message": f"Fabric f1 not found"
-    }
-    ```
-    """
-    fabric = session.get(FabricDbModel, fabric_name)
-    if not fabric:
-        detail = {"code": 404, "description": "", "message": f"Fabric {fabric_name} not found"}
-        raise HTTPException(status_code=404, detail=detail)
-    session.delete(fabric)
-    session.commit()
-    return {}
-
-
 @app.get(
     "/api/v1/manage/fabrics/{fabric_name}",
     response_model=FabricResponseModel,
